[Core][Graphics][DX12] Change `m_directQueueTimestampPeriod` to `double`
Adjusted `m_directQueueTimestampPeriod`'s type from `u64` to `double` in `Dx12GraphicsContext` for improved precision in timestamp calculations.

+++ Core/Src/Graphics/DirectX12/Dx12GraphicsContext.hpp
@@ -73,7 +73,7 @@ namespace KryneEngine
         bool m_enhancedBarriersEnabled = false;
 
         ComPtr<ID3D12QueryHeap> m_timestampQueryHeap { nullptr };
-        u64 m_directQueueTimestampPeriod = 0;
+        double m_directQueueTimestampPeriod = 0;
         u64 m_directQueueTimestampOffset = 0;
 
         void _CreateDevice(IDXGIFactory4* _factory4);
